Add reference test framework code, misc refactors.
