+++ eth-reference-tests/src/test/java/tech/pegasys/artemis/reference/PermutedIndexTestSuite.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package pegasys.artemis.reference;
+package tech.pegasys.artemis.reference;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -27,12 +27,14 @@ import java.util.stream.Stream;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.io.Resources;
 import org.apache.tuweni.junit.BouncyCastleExtension;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import tech.pegasys.artemis.datastructures.util.CrosslinkCommitteeUtil;
 
+@Disabled
 @ExtendWith(BouncyCastleExtension.class)
 public class PermutedIndexTestSuite {
