+++ eth-reference-tests/src/test/java/tech/pegasys/artemis/reference/framework/ReferenceTestHandler.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2019 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.artemis.reference.framework;
+
+/*
+ * The interface Eth2 spec reference test handlers implement.
+ */
+
+public interface ReferenceTestHandler {
+  // load test suites
+}
